+++ owncloudApp/src/main/java/com/owncloud/android/presentation/security/SecurityUtils.kt
@@ -22,7 +22,7 @@ package com.owncloud.android.presentation.security
 
 import android.os.SystemClock
 import com.owncloud.android.MainApp
-import com.owncloud.android.data.preferences.datasources.implementation.OCSharedPreferencesProvider
+import com.owncloud.android.data.providers.implementation.OCSharedPreferencesProvider
 
 const val PREFERENCE_LOCK_TIMEOUT = "lock_timeout"
 const val PREFERENCE_LAST_UNLOCK_TIMESTAMP = "last_unlock_timestamp"