+++ owncloudApp/src/main/java/com/owncloud/android/presentation/security/biometric/BiometricViewModel.kt
@@ -27,7 +27,7 @@ import android.security.keystore.KeyProperties
 import androidx.biometric.BiometricPrompt
 import androidx.lifecycle.ViewModel
 import com.owncloud.android.R
-import com.owncloud.android.data.preferences.datasources.SharedPreferencesProvider
+import com.owncloud.android.data.providers.SharedPreferencesProvider
 import com.owncloud.android.presentation.security.PREFERENCE_LAST_UNLOCK_TIMESTAMP
 import com.owncloud.android.presentation.security.passcode.PassCodeActivity
 import com.owncloud.android.providers.ContextProvider