+++ owncloudData/src/main/java/com/owncloud/android/data/providers/implementation/OCSharedPreferencesProvider.kt
@@ -17,12 +17,12 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package com.owncloud.android.data.preferences.datasources.implementation
+package com.owncloud.android.data.providers.implementation
 
 import android.content.Context
 import android.content.SharedPreferences
 import android.preference.PreferenceManager
-import com.owncloud.android.data.preferences.datasources.SharedPreferencesProvider
+import com.owncloud.android.data.providers.SharedPreferencesProvider
 
 class OCSharedPreferencesProvider(
     context: Context